Adding medications and editing them works just fine (including reminders handling)

+++ lib/widgets/medication/said_editable_med.dart
@@ -3,9 +3,12 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:said/config/color_constants.dart';
 import 'package:said/screens/medication/edit_medication_screen.dart';
+import 'package:said/services/medication_reminder_service.dart';
 import 'package:said/services/medication_service.dart';
 import 'package:said/services/models/medication.dart';
+import 'package:said/services/models/medication_reminder.dart';
 import 'package:said/services/models/user.dart';
+import 'package:said/services/utils/medication_reminder_utils.dart';
 import 'package:said/widgets/buttons/said_fab.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
@@ -14,15 +17,21 @@ class SaidEditableMed extends StatelessWidget {
       {super.key,
       required this.authenticatedUser,
       required this.medication,
-      required this.onRefreshUser});
+      required this.onRefreshScreen});
 
   final User authenticatedUser;
   final Medication medication;
-  final Function(User) onRefreshUser;
+  final Function(User) onRefreshScreen;
 
   Future<void> _deleteMedication(BuildContext context) async {
-    var response =
-    await MedicationService.deleteMedication(medication.id!);
+    // first things first, delete associated medication reminders:
+    bool success = await deleteAssociatedMedicationReminders(medication);
+    if (!success) {
+      return;
+    }
+
+    // second, delete medication itself:
+    var response = await MedicationService.deleteMedication(medication.id!);
 
     if (response.statusCode != 200) {
       var body = jsonDecode(response.body);
@@ -31,13 +40,13 @@ class SaidEditableMed extends StatelessWidget {
       // show snackbar:
       final snackBar = SnackBar(
         content:
-        Text('${AppLocalizations.of(context).changesSavedError}: $errMsg'),
+            Text('${AppLocalizations.of(context).changesSavedError}: $errMsg'),
         backgroundColor: Colors.red,
         behavior: SnackBarBehavior.floating,
       );
       ScaffoldMessenger.of(context).showSnackBar(snackBar);
     } else {
-      onRefreshUser(authenticatedUser);
+      onRefreshScreen(authenticatedUser);
 
       // show snackbar:
       final snackBar = SnackBar(
@@ -68,7 +77,7 @@ class SaidEditableMed extends StatelessWidget {
                           decoration: BoxDecoration(
                               border: Border.all(color: Colors.grey),
                               borderRadius:
-                              const BorderRadius.all(Radius.circular(16))),
+                                  const BorderRadius.all(Radius.circular(16))),
                           child: const Padding(
                               padding: EdgeInsets.all(4.0),
                               child: Icon(Icons.medical_services_rounded,
@@ -100,7 +109,7 @@ class SaidEditableMed extends StatelessWidget {
                       icon: const Icon(Icons.edit),
                       linkTo: EditMedicationScreen(
                         authenticatedUser: authenticatedUser,
-                        onRefreshScreen: onRefreshUser,
+                        onRefreshScreen: onRefreshScreen,
                         medication: medication,
                       )),
                   const Padding(padding: EdgeInsets.all(4)),
